model/requirements.txt
torch>=1.7.0
torchvision>=0.8.1
opencv-python>=4.6.0
PyYAML>=5.3.1
numpy>=1.19.5
Pillow>=7.1.2
tqdm>=4.64.0
ultralytics
albumentations
mlflow
scipy
pytest
